Migrated some code to util class
