+++ scala-debugger-macros/src/main/scala/org/scaladebugger/macros/freeze/FreezableMacro.scala
@@ -1,23 +1,25 @@
 package org.scaladebugger.macros.freeze
 
 import scala.language.experimental.macros
-import scala.annotation.tailrec
 import scala.reflect.macros.whitebox
 import macrocompat.bundle
+import org.scaladebugger.macros.MacroUtils
 
 @bundle class FreezableMacro(val c: whitebox.Context) {
   import c.universe._
+
+  val M = new MacroUtils[c.type](c)
 
   // NOTE: Stuck with c.Expr[Any] instead of c.Tree for now;
   //       blows up in Scala 2.10 even with @bundle otherwise
   def impl(annottees: c.Expr[Any]*): c.Expr[Any] = {
     val (annottee, expandees) = annottees.map(_.tree) match {
-      case (classDef: ClassDef) :: (moduleDef: ModuleDef) :: Nil if isTrait(classDef) =>
+      case (classDef: ClassDef) :: (moduleDef: ModuleDef) :: Nil if M.isTrait(classDef) =>
         //println("INPUT CLASS AND MODULE :: " + classDef + " :: " + moduleDef)
         val results = processTraitAndObj(classDef, moduleDef)
         //println("RESULTS :: " + results)
         (EmptyTree, results)
-      case (classDef: ClassDef) :: Nil if isTrait(classDef) =>
+      case (classDef: ClassDef) :: Nil if M.isTrait(classDef) =>
         //println("INPUT CLASS ONLY :: " + classDef)
         val results = processTrait(classDef)
         //println("RESULTS :: " + results)
@@ -33,113 +35,6 @@ import macrocompat.bundle
     c.Expr[Any](
       Block(outputs, Literal(Constant(())))
     )
-  }
-
-  private def containsAnnotation(d: DefDef, aName: String): Boolean =
-    findAnnotation(d, aName).nonEmpty
-
-  private def findAnnotation(d: DefDef, aName: String): Option[Tree] =
-    d.mods.annotations.find(a => isAnnotation(a, aName))
-
-  private def isAnnotation(t: Tree, aName: String): Boolean = t match {
-    case Apply(f, args) => f match {
-      case Select(qual, name) => qual match {
-        case New(tpt) => tpt match {
-          case Ident(iName) => iName.decodedName.toString == aName
-          case _ => false
-        }
-        case _ => false
-      }
-    }
-    case _ => false
-  }
-
-  private def isTrait(classDef: ClassDef): Boolean = {
-    try {
-      val q"""
-          $mods trait $tpname[..$tparams] extends {
-            ..$earlydefns
-          } with ..$parents { $self =>
-            ..$body
-          }
-        """ = classDef
-      true
-    } catch {
-      case _: Throwable => false
-    }
-  }
-
-  private def extractInheritedMethods(parents: List[Type]): List[MethodSymbol] = {
-    val anyTypeSymbol = typeOf[Any].typeSymbol
-    val objectTypeSymbol = typeOf[Object].typeSymbol
-
-    def stripDuplicateMethods(methods: List[MethodSymbol]): List[MethodSymbol] = {
-      val overrides = methods.flatMap(_.overrides)
-      methods.filterNot(overrides.contains)
-    }
-
-    @tailrec def extract(types: List[Type], methods: List[MethodSymbol]): List[MethodSymbol] = {
-      val parentMethods = types.flatMap(pt => {
-        pt.members
-          .filter(_.isMethod)
-          .filter(_.owner == pt.typeSymbol)
-          .map(_.asMethod)
-      })
-      val parentTypes = types
-        .flatMap(t => {
-          t.baseClasses.filterNot(c =>
-            c == t.typeSymbol ||
-            c == anyTypeSymbol ||
-            c == objectTypeSymbol
-          )
-        })
-        .filter(_.isClass)
-        .map(_.asClass.toType)
-
-      val allMethods = stripDuplicateMethods(methods ++ parentMethods)
-
-      if (parentTypes.nonEmpty) extract(parentTypes, allMethods)
-      else allMethods
-    }
-
-    extract(parents, Nil)
-  }
-
-  private def extractParentTypes(parents: List[Tree]): List[Option[Type]] = {
-    parents match {
-      case p: List[Tree] => p.map(extractParentType)
-    }
-  }
-
-  private def extractParentType(parent: Tree): Option[Type] = {
-    def isBaseType(n1: Name, n2: Name = null): Boolean = {
-      val name = if (n2 != null) {
-        n1.toString + "." + n2.toString
-      } else {
-        n1.toString
-      }
-
-      val baseTypeNames = Seq(
-        "Any", "scala.Any", "AnyRef", "scala.AnyRef",
-        "AnyVal", "scala.AnyVal", "java.lang.Object"
-      )
-
-      baseTypeNames.contains(name)
-    }
-
-    parent match {
-      case Ident(name) if !isBaseType(name) =>
-        val typeName = name.toTypeName
-        val typedTree = scala.util.Try(c.typecheck(q"null.asInstanceOf[$typeName]"))
-        typedTree.toOption.flatMap(tt => Option(tt.tpe))
-      case Select(Ident(name), typeName) if !isBaseType(name, typeName) =>
-        val t1 = name.toTermName
-        val t2 = typeName.toTypeName
-        val typedTree = scala.util.Try(c.typecheck(q"null.asInstanceOf[$t1.$t2]"))
-        typedTree.toOption.flatMap(tt => Option(tt.tpe))
-      case a =>
-        None
-    }
   }
 
   private def processTrait(classDef: ClassDef): List[Tree] = {
@@ -159,9 +54,7 @@ import macrocompat.bundle
 
     // Generate a new class containing the helper methods and
     // object containing "Frozen" class representation
-    val moduleDef: ModuleDef = q"object $oName {}" match {
-      case m: ModuleDef => m
-    }
+    val moduleDef = M.newEmptyObject(traitTypeName.toString)
 
     processTraitAndObj(classDef, moduleDef)
   }
@@ -178,155 +71,16 @@ import macrocompat.bundle
       }
     """ = classDef
 
-    val inheritedMethods = extractInheritedMethods(
-      extractParentTypes(parents.asInstanceOf[List[Tree]]).flatten
+    val inheritedMethods = M.extractInheritedMethods(
+      M.extractParentTypes(parents.asInstanceOf[List[Tree]]).flatten
     )
 
     // Represents name of variable passed to freeze method
-    val freezeObjName = TermName("valueToFreeze")
 
-    val freezeTrees = {
-      val cParams = collection.mutable.ArrayBuffer[Tree]()
-      val cArgs = collection.mutable.ArrayBuffer[Tree]()
-      val bodyTrees: List[Tree] = body match {
-        case l: List[Tree] => l
-      }
-      val internalMethods = bodyTrees.collect { case d: DefDef => d }
-
-      // TODO: Also process inherited methods that are not covered
-      //       by the tree methods
-      // TODO: CHIP CHIP CHIP CHIP CHIP CHIP CHIP CHIP CHIP CHIP CHIP
-      val internals = internalMethods.map {
-        case d: DefDef if containsAnnotation(d, "CanFreeze") =>
-          val q"$mods def $tname[..$tparams](...$paramss): $tpt = $expr" = d
-
-          val paramTrees: List[List[ValDef]] = paramss match {
-            case l: List[List[ValDef]] => l
-          }
-          if (paramTrees.nonEmpty) c.abort(
-            c.enclosingPosition,
-            s"Unable to freeze $tname! Must have zero arguments!"
-          )
-
-          val name = TermName(s"$$$tname")
-          cParams.append(q"""
-            private val $name: scala.util.Try[$tpt]
-          """)
-
-          // TODO: Remove runtime checks by getting better understanding of
-          //       return type, if it contains a freeze method, etc.
-          findAnnotation(d, "CanFreeze").get match {
-            case q"new CanFreeze()" =>
-              cArgs.append(q"""
-                scala.util.Try.apply($freezeObjName.$tname)
-                  .asInstanceOf[scala.util.Try[$tpt]]
-              """)
-            case q"new CanFreeze($returnType)" =>
-              // TODO: Find better way to match against the input
-              import org.scaladebugger.macros.freeze.CanFreeze.ReturnType
-              val r = ReturnType.withName(returnType.toString().split('.').last)
-
-              r match {
-                case ReturnType.Normal =>
-                  cArgs.append(q"""
-                    scala.util.Try.apply($freezeObjName.$tname)
-                      .asInstanceOf[scala.util.Try[$tpt]]
-                  """)
-                case ReturnType.FreezeObject =>
-                  cArgs.append(q"""
-                    scala.util.Try.apply($freezeObjName.$tname).map(_.freeze())
-                      .asInstanceOf[scala.util.Try[$tpt]]
-                  """)
-                case ReturnType.FreezeCollection =>
-                  cArgs.append(q"""
-                    scala.util.Try.apply($freezeObjName.$tname).map(_.map(_.freeze()))
-                      .asInstanceOf[scala.util.Try[$tpt]]
-                  """)
-                case ReturnType.FreezeOption =>
-                  cArgs.append(q"""
-                    scala.util.Try.apply($freezeObjName.$tname).map(_.map(_.freeze()))
-                      .asInstanceOf[scala.util.Try[$tpt]]
-                  """)
-                case ReturnType.FreezeEither =>
-                  cArgs.append(q"""
-                    scala.util.Try.apply($freezeObjName.$tname).map(r => r match {
-                      case scala.util.Left(o) => scala.util.Left(o.freeze())
-                      case scala.util.Right(o) => scala.util.Right(o.freeze())
-                    }).asInstanceOf[scala.util.Try[$tpt]]
-                  """)
-              }
-          }
-
-          // Add override if not already there
-          val oldMods: Modifiers = mods match {
-            case m: Modifiers => m
-          }
-
-          // TODO: Is there a way to filter out just the DEFERRED flag?
-          val newFlags = Flag.OVERRIDE |
-            (if (oldMods.hasFlag(Flag.FINAL)) Flag.FINAL else NoFlags)
-
-          val newMods = c.universe.Modifiers(
-            flags = newFlags,
-            privateWithin = oldMods.privateWithin,
-            annotations = oldMods.annotations.filterNot(a => {
-              isAnnotation(a, "CanFreeze") || isAnnotation(a, "CannotFreeze")
-            })
-          )
-
-          q"""
-            $newMods def $tname[..$tparams](...$paramss): $tpt = {
-              this.$name.get
-            }
-          """
-        case d: DefDef if containsAnnotation(d, "CannotFreeze") =>
-          val q"$mods def $tname[..$tparams](...$paramss): $tpt = $expr" = d
-
-          // Add override if not already there
-          val oldMods: Modifiers = mods match {
-            case m: Modifiers => m
-          }
-
-          // TODO: Is there a way to filter out just the DEFERRED flag?
-          val newFlags = Flag.OVERRIDE |
-            (if (oldMods.hasFlag(Flag.FINAL)) Flag.FINAL else NoFlags)
-
-          val newMods = c.universe.Modifiers(
-            flags = newFlags,
-            privateWithin = oldMods.privateWithin,
-            annotations = oldMods.annotations.filterNot(a => {
-              isAnnotation(a, "CanFreeze") || isAnnotation(a, "CannotFreeze")
-            })
-          )
-
-          q"""
-            $newMods def $tname[..$tparams](...$paramss): $tpt = {
-              throw new IllegalStateException("Method not frozen!")
-            }
-          """
-        case d: DefDef =>
-          val q"$mods def $tname[..$tparams](...$paramss): $tpt = $expr" = d
-          val exprTree: Tree = expr match {
-            case t: Tree => t
-          }
-          if (exprTree.isEmpty) c.abort(
-            c.enclosingPosition,
-            s"$tname has no implementation and is not marked as freezable or unfreezable!"
-          )
-          null
-      }.filterNot(_ == null)
-
-      val klass = q"""
-        final class Frozen(..$cParams) extends $tpname with java.io.Serializable { $self =>
-          ..$internals
-        }
-      """
-      val method = q"""
-        def freeze($freezeObjName: $tpname): $tpname = new Frozen(..$cArgs)
-      """
-
-      List(klass, method)
-    }
+    val freezeTrees = generateTrees(
+      body match { case l: List[Tree] => l },
+      tpname match { case t: TypeName => t }
+    )
 
     // TODO: Inject freeze method directly into trait rather than
     //       relying on object implicit
@@ -352,7 +106,9 @@ import macrocompat.bundle
         }
       """
 
-      tree match { case c: ClassDef => c }
+      tree match {
+        case c: ClassDef => c
+      }
     }
 
     val newModuleDef: ModuleDef = {
@@ -364,23 +120,168 @@ import macrocompat.bundle
         }
       """ = moduleDef
 
-      val oldBody: List[Tree] = body match {
-        case l: List[Tree] => l
-      }
-      val newBody = oldBody ++ freezeTrees
-
       val newObjDef = q"""
-          $mods object $tname extends {
-            ..$earlydefns
-          } with ..$parents { $self =>
-            ..$newBody
-          }
-        """
+        $mods object $tname extends {
+          ..$earlydefns
+        } with ..$parents { $self =>
+          ..$body
+          ..$freezeTrees
+        }
+      """
 
       newObjDef match { case m: ModuleDef => m }
     }
 
     List(newClassDef, newModuleDef)
+  }
+
+  private def generateTrees(body: List[Tree], traitTypeName: TypeName): List[Tree] = {
+    val freezeObjName = TermName("valueToFreeze")
+    val cParams = collection.mutable.ArrayBuffer[Tree]()
+    val cArgs = collection.mutable.ArrayBuffer[Tree]()
+    val bodyTrees: List[Tree] = body match {
+      case l: List[Tree] => l
+    }
+    val internalMethods = bodyTrees.collect { case d: DefDef => d }
+
+    // TODO: Also process inherited methods that are not covered
+    //       by the tree methods
+    // TODO: CHIP CHIP CHIP CHIP CHIP CHIP CHIP CHIP CHIP CHIP CHIP
+    //
+    // See https://github.com/paulbutcher/ScalaMock/blob/master/shared/src/main/scala/org/scalamock/clazz/MockMaker.scala
+    // for example of loading inherited methods in method symbols not found in object
+    val internals = internalMethods.map {
+      case d: DefDef if M.containsAnnotation(d, "CanFreeze") =>
+        val q"$mods def $tname[..$tparams](...$paramss): $tpt = $expr" = d
+
+        val paramTrees: List[List[ValDef]] = paramss match {
+          case l: List[List[ValDef]] => l
+        }
+        if (paramTrees.nonEmpty) c.abort(
+          c.enclosingPosition,
+          s"Unable to freeze $tname! Must have zero arguments!"
+        )
+
+        val name = TermName(s"$$$tname")
+        cParams.append(q"""
+            private val $name: scala.util.Try[$tpt]
+          """)
+
+        // TODO: Remove runtime checks by getting better understanding of
+        //       return type, if it contains a freeze method, etc.
+        M.findAnnotation(d, "CanFreeze").get match {
+          case q"new CanFreeze()" =>
+            cArgs.append(q"""
+                scala.util.Try.apply($freezeObjName.$tname)
+                  .asInstanceOf[scala.util.Try[$tpt]]
+              """)
+          case q"new CanFreeze($returnType)" =>
+            // TODO: Find better way to match against the input
+            import org.scaladebugger.macros.freeze.CanFreeze.ReturnType
+            val r = ReturnType.withName(returnType.toString().split('.').last)
+
+            r match {
+              case ReturnType.Normal =>
+                cArgs.append(q"""
+                    scala.util.Try.apply($freezeObjName.$tname)
+                      .asInstanceOf[scala.util.Try[$tpt]]
+                  """)
+              case ReturnType.FreezeObject =>
+                cArgs.append(q"""
+                    scala.util.Try.apply($freezeObjName.$tname).map(_.freeze())
+                      .asInstanceOf[scala.util.Try[$tpt]]
+                  """)
+              case ReturnType.FreezeCollection =>
+                cArgs.append(q"""
+                    scala.util.Try.apply($freezeObjName.$tname).map(_.map(_.freeze()))
+                      .asInstanceOf[scala.util.Try[$tpt]]
+                  """)
+              case ReturnType.FreezeOption =>
+                cArgs.append(q"""
+                    scala.util.Try.apply($freezeObjName.$tname).map(_.map(_.freeze()))
+                      .asInstanceOf[scala.util.Try[$tpt]]
+                  """)
+              case ReturnType.FreezeEither =>
+                cArgs.append(q"""
+                    scala.util.Try.apply($freezeObjName.$tname).map(r => r match {
+                      case scala.util.Left(o) => scala.util.Left(o.freeze())
+                      case scala.util.Right(o) => scala.util.Right(o.freeze())
+                    }).asInstanceOf[scala.util.Try[$tpt]]
+                  """)
+            }
+        }
+
+        // Add override if not already there
+        val oldMods: Modifiers = mods match {
+          case m: Modifiers => m
+        }
+
+        // TODO: Is there a way to filter out just the DEFERRED flag?
+        val newFlags = Flag.OVERRIDE |
+          (if (oldMods.hasFlag(Flag.FINAL)) Flag.FINAL else NoFlags)
+
+        val newMods = c.universe.Modifiers(
+          flags = newFlags,
+          privateWithin = oldMods.privateWithin,
+          annotations = oldMods.annotations.filterNot(a => {
+            M.isAnnotation(a, "CanFreeze") ||
+              M.isAnnotation(a, "CannotFreeze")
+          })
+        )
+
+        q"""
+            $newMods def $tname[..$tparams](...$paramss): $tpt = {
+              this.$name.get
+            }
+          """
+      case d: DefDef if M.containsAnnotation(d, "CannotFreeze") =>
+        val q"$mods def $tname[..$tparams](...$paramss): $tpt = $expr" = d
+
+        // Add override if not already there
+        val oldMods: Modifiers = mods match {
+          case m: Modifiers => m
+        }
+
+        // TODO: Is there a way to filter out just the DEFERRED flag?
+        val newFlags = Flag.OVERRIDE |
+          (if (oldMods.hasFlag(Flag.FINAL)) Flag.FINAL else NoFlags)
+
+        val newMods = c.universe.Modifiers(
+          flags = newFlags,
+          privateWithin = oldMods.privateWithin,
+          annotations = oldMods.annotations.filterNot(a => {
+            M.isAnnotation(a, "CanFreeze") ||
+              M.isAnnotation(a, "CannotFreeze")
+          })
+        )
+
+        q"""
+            $newMods def $tname[..$tparams](...$paramss): $tpt = {
+              throw new IllegalStateException("Method not frozen!")
+            }
+          """
+      case d: DefDef =>
+        val q"$mods def $tname[..$tparams](...$paramss): $tpt = $expr" = d
+        val exprTree: Tree = expr match {
+          case t: Tree => t
+        }
+        if (exprTree.isEmpty) c.abort(
+          c.enclosingPosition,
+          s"$tname has no implementation and is not marked as freezable or unfreezable!"
+        )
+        null
+    }.filterNot(_ == null)
+
+    val klass = q"""
+        final class Frozen(..$cParams) extends $traitTypeName with java.io.Serializable {
+          ..$internals
+        }
+      """
+    val method = q"""
+        def freeze($freezeObjName: $traitTypeName): $traitTypeName = new Frozen(..$cArgs)
+      """
+
+    List(klass, method)
   }
 }
